getChunk now has flag bit
The getChunk function now treats the most significant bit as a flag for
whether or not it is a specifically placed chunk. It will be set to true
if it is a specifically placed chunk and false if it is part of a
region.

Changed the design of the test world slightly and added a chunk.