Additional Features, Bug Fixes
- Fixed entering battle flashing effect not playing after escaping from
a battle.
- The text engine now handles the input for rolling text to fix freezing
issues.
- Changed the order of input in relation to the game state to ensure the
text inputs and game inputs do not interfere with each other.
- Instead of increasing the last Boss's health when the blade is not
equipped, the Boss's attack is now increased.
- Player can no longer walk outside map. Renders rocks outside map area.
- Mini map now centers on the player when upon opening.
- Fixed NPC breathing.
- Fixed NPC's all looked the same.
- Boss now has max defense and special defense if blade is not equipped.
- Enemies now get progressively bigger by area instead of having all
sizes. Originally the size of the enemy was supposed to be an indicator
of how difficult the enemy would be within that area, but it didn't end
up being the case so smaller enemies could be stronger than larger
enemies. Now headsize is purely determined by level.
- Changed the description for the Amulet of Victory from "A SYMBOL OF
VICTORY" to "THANK YOU FOR PLAYING!" to better convey the end of the
game.
- Upload message now says "DOWN" key instead up "UP" to work with
updated Arduboy2 library.
- Added B button map shortcut when not investigating an object or NPC.
- Fixed bosses giving amulets more than once.
- Text no longer runs over the intended buffer amount.
- Fixed returning to menu from in game putting cursor outside screen.
- Cleaned up some commented out code.